Reorganize file structure to support pyro-api dispatch
