+++ pyro/contrib/__init__.py
@@ -9,7 +9,7 @@ Contributed Code
     This code makes no guarantee about maintaining backwards compatibility.
 """
 
-from pyro.contrib import autoname, bnn, easyguide, epidemiology, forecast, gp, oed, tracking
+from pyro.contrib import autoname, bnn, easyguide, epidemiology, forecast, funsor, gp, oed, tracking
 
 __all__ = [
     "autoname",
@@ -17,6 +17,7 @@ __all__ = [
     "easyguide",
     "epidemiology",
     "forecast",
+    "funsor",
     "gp",
     "oed",
     "tracking",
